# internal/api/model_result.go
/*
 * Bactic
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api




type Result struct {

	Id int64 `json:"id"`

	HeatId int64 `json:"heat_id"`

	AthleteId int64 `json:"athlete_id"`

	Place int32 `json:"place"`

	Quantity float32 `json:"quantity"`

	WindMs float32 `json:"wind_ms,omitempty"`

	Stage int32 `json:"stage,omitempty"`

	Team string `json:"team,omitempty"`
}

// AssertResultRequired checks if the required fields are not zero-ed
func AssertResultRequired(obj Result) error {
	elements := map[string]interface{}{
		"id": obj.Id,
		"heat_id": obj.HeatId,
		"athlete_id": obj.AthleteId,
		"place": obj.Place,
		"quantity": obj.Quantity,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertResultConstraints checks if the values respects the defined constraints
func AssertResultConstraints(obj Result) error {
	return nil
}