# internal/api/model_team_summary.go
/*
 * Bactic
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api




type TeamSummary struct {

	Name string `json:"name"`

	Id int64 `json:"id"`

	Zscore float32 `json:"zscore,omitempty"`

	Athletes []Athlete `json:"athletes,omitempty"`
}

// AssertTeamSummaryRequired checks if the required fields are not zero-ed
func AssertTeamSummaryRequired(obj TeamSummary) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"id": obj.Id,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	for _, el := range obj.Athletes {
		if err := AssertAthleteRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertTeamSummaryConstraints checks if the values respects the defined constraints
func AssertTeamSummaryConstraints(obj TeamSummary) error {
	return nil
}